refactor(export-image): move dimension label functions
